[test] SwiftFixIt: Factor out tests into separate files
